Moved old PublicAddress tests into EthAddress tests. Refactored to match new component structure. Updated snapshots.
